move FixtureDataTest to own file
