+++ corehq/apps/fixtures/tests/__init__.py
@@ -1,1 +1,2 @@
 from .tests import *
+from .test_fixture_data import *
